Delete obsolete `UserModifiesRGBLayers` class
With this commit all functionalities of the rgb composites
no longer work correctly.
The complete workflow with RGB Composites has to be revised
so that it can be managed by the new Layer Manager system.
In the new workflow of RGB Composites, a User Behaviour class
is no longer needed, so it can be deleted completely.