Enhance Layer Manager UI / PieDial interaction
Directly react to a mouse button press in the PieDial, not only when
releasing. This just feels more natural.  To allow this, some more mouse
Events must be "accepted" by the PieDialWidget to prevent:
- selection of current layer with the first click in the PieDial
- selection of layers when moving the mouse out of the PieDial while
  mouse button is pressed
Disadvantage (which can be made an advantage): Starting layer resorting
by drag and drop can not be initated by clicking on the PieDial of a
layer entry. But: this allows for implementing the manipulation of the
opacity by moving the mouse while a button is pressed.