Implement layer reordering via z order in scene graph
Introduce new signal 'didReorderLayers' in LayerModel. This signal is
connected to a slot in 'scene_graph.py', allowing changes to the layer
order to be propagated to SceneGraphManager.  The latter then adjusts
the transformation of the layer nodes accordingly, so that the plane of
the lower layers (which actually have a higher row number in the
LayerModel) has a lower (more negative) z-coordinate (z grows towards
the viewer).