Introduce signal LayerModel.didUpdateLayers
which is emitted whenever any change in the available datasets (i.e.
changing any timeline of any layer) occurs. Currently this can only
happen when datasets are *added* (removal of datasets is not yet
possible).

This signal is - of course - intended to be received by the TimeManager,
which is supposed to update its state accordingly.