Revise Time{Manager,Transformer} step control
There is the concept of the 'tick' (as in a clock), which triggers a
'step' on the timeline of the timebase (i.e., driving layer).  This
invokes the TimeMatcher to find a dataset for each layer that best
matches the then current time, based on its active time matching policy.