Move driving layer selection convenience function(s) to QMLLayerManager
Convenience functions are supposed to aid in selecting a driving layer
according to some criterion.  The so far only comfort function "Most
frequent" offers the selection according to the level with the finest
time resolution.

These convenience functions used to reside in the now deprecated
DataLayerCollection and are moved to QMLLayerManager, as the user
will also select these functions from within the QML part of the UI.