Hook up SceneGraphManager to animate through datasets
To that end a new signal 'LayerModel.didActivateProductDataset' is
connected to 'SceneGraphManager.change_dataset_visibile'.

In the LayerModel 'active' datasets (ProductDataset.is_active == True)
are those, which should show up at the current state of a running
animation/the currently active display time (when stepping through the
timeline manually).  Therefore, changes to the 'is_activate' status of
the datasets must be passed on to the scene graph so that the 'visible'
status of the corresponding dataset nodes is changed.