Provide API to get active datasets from LayerModel
... needed in the following to retrieve data at point probe location
(and later for probing region) from them.

The LayerModel provides functionality to determine the (LayerItem of
the) top layer with probeable data, whereas the LayerItem provides the
functionality to retrieve the (first of its) active ProductDatasets.

Probeable data is (currently) of kind IMAGE or COMPOSITE, i.e. single
channel raster data. Probing POINTS data may be added in the future.