Connect signals 'pointProbeChanged' to 'on_point_probe_set'
Connect graphManger signal to LayerModel slot yielding an update to the
layer manager 'probe_value' column if a point probe is set.

It is not sufficient to update probe values for active datasets only, it
must be done for all layers: In a previous time step, a layer may have
had an active dataset, but not in the current one. In this case, the
probe value for that layer would remain at the value for the previous
time step and not be updated to 'n/a'.

Ensure proper updates by iterating over all 'probeable' layers to update
their probe values.