Remove unused arguments from signal 'newProbePolygon'
Originally the 'newProbePolygon' signal carries a UUID, as well as an
array of points (describing the polygon). Let it only carry the array of
points.
This change is made due to the UUID argument of the signal being unused
in the connected slot 'Main.update_probe_polygon'. The slot signature is
also changed to reflect that.