Remove splat operator from one method signature
and calls of it: ProbeGraphManager.set_default_layer_selections(...)

Prior to this commit lists of uuids are flattened into separate
arguments using fun(*arg) even though varargs are not necessary in the
methods themselves, as only a list is passed as argument anyways.

Simplify the interface by removing this superflous argument conversion
and always pass/take a list.