Implement switching automatic/manual update of probe graphs
Let ProbeGraphManager manage two modes: updating the probe graphs on
every time step or due to other events on the one hand or only manually,
as triggered by the user on the other hand. The modes can be switched by
the user by (un)checking the  "Auto Update" check box.