Provide LayerItem with excerpt of dataset info
Copy that information from the generating dataset when the LayerItem is
created via add_dataset.  The method extract_layer_info() holds a list
of keys for which data is to be copied.  This list is to be extended
according to future needs.  The method handles missing keys in the input
dataset info dict gracefully and ignores them.