Remove references to 'Document' from probes.py (WIP)
in particular in the __init__ method of
ProbeGraphManager/ProbeGraphDisplay, i.e., do not initialize
self.document any more in these classes.

This is in preparation to use the LayerModel instead of the Document to
retrieve layers and datasets remove 'document' from the class
*interfaces*.  The implementation still contains references to
self.document,

 THUS THIS IS A NOT 'COMPILING' CODE STATE.

TODO: replace occurences of self.document by appropriate calls to the to
be introduced LayerModel.