Fix ProbeGraphDisplay broken class variables
It looks like that in the legacy implementation it was not understood
how member variables are declared/defined in Python and some were
wrongly declared as class variables. Thus all instances of
ProbeGraphDisplay shared them - strange enough, that this ever worked
correctly (if it did).

Move the variable declarations into ProbeGraphDisplay.__init__(...) to
fix this.