Keep timeline driving layer selection in sync with layer order
When layers are re-orderd in the Layer Manager the combo box to select
the driving layer in the Timeline Manager should update to display
layers in the same order but keep the current driving layer active as
such.

Layer resorting in layer manager updates up to the order layers are
presented in the Timebase selection ComboBox.