Automate Layer Manager columns resizing
by applying settings to the header which let it automatically adjust
column widths to the content width for the price of not allowing column
resizing by the user.

TODO: The automatic column resizing may be too restrictive, users may
want to resize columns interactively. Revise the changes of this commit
if so.