Make 'most frequent' driving layer selection more robust
get_most_frequent_data_layer_index() will not return a better layer if
there is a layer that has only one (or zero) dataset(s) - this (or one
of them) is currently selected.

Make sure that such layers are considered with the lowest priority, so
that a layer with two or more data sets is always preferred.