Enable Content subclasses to get all metadatabase table column names
To that end use 'single table inheritance' in metadatabase.

The problem to be solved: The auxiliary class function
'Content._separate_fields_and_keys()' is used in 'Content.from_info()'
to create a 'Product' from an info dict. In doing so, it needs to know
all column names of the database table for the *concrete* Content object
which is of a derived type (ContentImage, ContentLines,
ContentUnstructuredPoints). With the previous choice of SQLAlchemy's
style of database inheritance - "joined" - the column names of the base
class table can not be determined (or only with complicated
work-arounds) by the derived Content classes, but with "single" table
inheritance it becomes easy.

Thus: Switch to single table inheritance which requires to resolve one
column conflict: both ContentLines and ContentUnstructuredPoints have a
field `n_dimensions`, make them share an according table column.

See:
- https://docs.sqlalchemy.org/en/13/orm/inheritance.html#single-table-inheritance
- https://docs.sqlalchemy.org/en/13/orm/extensions/declarative/inheritance.html#single-table-inheritance