Remove prefix 'n_' form metadatabase columns
n_{rows,cols,levels} -> {rows,cols,levels}

In commit

d84e2a0c ("Add Content subclasses to metadatabase", 2021-10-11)

the metadatabase tables/variables `rows`, `cols`, `levels`, which denote
counts, were prefixed with `n_` to make this clearer. This now turned
out to be problematic, since at other locations they are still in use
without the prefix (automatic refactoring didn't catch these cases and
also with looking at the code it is sometimes not clear, whether
variables of the same name are related or not).

Thus undo this change.

TODO: To keep inner consistency in the metadatabase also the following
variables could be renamed (which is not done now because of reasons):
 - n_attributes
 - n_dimensions
 - n_lines
 - n_points