Fix invalid timeline index after dataset removal
When timelines change the TimeTransformerPolicy tries to find the
time key in the newly active timeline closest to the previous simulation
time.

The current approach is to take that time via its index - stored
in `_driving_idx` - from the timeline of the previous driving layer.
This doesn't cover the case, that the timeline of the driving layer has
changed under the hood e.g. due to added or removed datasets - the index
is already invalid then.  Especially in the latter case the said index
may be out of bounds, which shows up as an Exception.

The safer way is to store the simulation time in `_curr_t_sim` whenever
`_driving_idx` or `driving_layer` changes (in this use case, storing
redundant data is helpful but requires proper maintenance).