Fix missing indirect update of RGBs with Algebraics input
RGB composites which take input from an algebraic do not get informed,
when new times for the inputs of the algebraic become available and the
algebraic is updated (by clicking "Update") for them. As a result no
RGBs for the new times appear in the timeline even if all their input
requirements are satisfied.

Fix this by adding an according trigger.