Eliminate RuntimeError regarding QTreeView instance
Undo the deletion of self.ui.layersPane
(from 89044aaf ("Hide legacy layersPane from MainWindow", 2022-02-21))
for now, because if it is deleted all remaining references to it will
result (of course) in annoying error messages for the old layers_tree
owned by it and deleted together with it:

"RuntimeError: wrapped C/C++ object of type QTreeView has been deleted"

The complete removal of the old layersPane and its layer_tree is the
task of issue #89.

This removes error messages mentioned in #113 and #131 which are not
related to their very topic.