Fix image/movie export
Re-enable to export image(s) or movie of the current animation by
porting the existing export functionality to the new
LayerManager/AnimationController/TimeManager concept.
The frames of the current animation are defined by the times of the
driving layer.

Implementation details:
- provide low level TimeManager API to query for data about the current
  timeline and its datasets
- provide more abstract AnimationController API to query for information
  about frames of the current animation
- port the image export functionalities to use the new
  AnimationController API

Fixes #115

This commit also closes #105, as it makes the according feature available
again.