Enable keyboard toggle of visibility of selected layer
The keyboard shortcut 'V' used to toggle the visibility of a legacy
layer (which is now a product dataset), but that of course doesn't make
sense with the new layer management. The keyboard shortcut is still
displayed in the Edit menu but does nothing.

Of course, the user expects to be able to toggle the visibility of a
layer by hitting 'V', as is possible by clicking on the visibility
control in the layer manager.

Add according functionality to make the shortcut work for layer
visibility toggling.

Concerns #125