Avoid abort of loading due to another exception
Similar to the previous commit introduce a check to avoid an
uncaught exception in the following case: an incomplete file set (e.g.
SEVIRI HRIT: for a start time only EPI or EPI+PRO but not a data stripe
or a data stripe with at least one of EPI, PRO missing) is picked up by
the catalogue in "Auto-Update" mode. This can easily happen for SEVIRI
since the HRIT files come in piece by piece.