Fix edited colormap not used by all datasets
In the current state, the representation of the layer is not updated
when changes are made, e.g. when the colour chart for the layer is
changed. Instead, the new value for the representation attribute is
stored in a new variable. This variable has no relation to the
representation of the layer. So nothing has changed in the view of
the layer, but the changes are passed on to the Scenegraph, where
the changes are adapted to the currently available visual nodes of
the `ProductDataset`. However, these changes could not be applied
to newly loaded datasets, because the visualisation of the layer was
still unchanged.

With commit, the behaviour is fixed and works as intended.

Fixes #114, #118, #119