Port layer visibility toggling to (multi-)selection
There is a difference between selected item(s) and current item in a
QTreeView: the current item is not necessarily selected and when
multi-selection is enabled (which may become the case in the future) at
most one of several selected items can be current. Therefore the more
correct, robust and future-proof implementation does not rely on the
currentIndex() but on the selectedIndexes().

Implement this enhancement.