Prevent mock data script from generating start times in the future
if the given --time_delay is shorter than the nominal repeat cycle of
the mocked data (SEVIRI: full disk every 15 minutes, RSS every 5
minutes; FCI: new file every 15 seconds) the script produce start times
in the filename which rush ahead into the future which does not create a
valid test situation.

To cope for this restrict the start time in the filename of a mock data
file to be not later than "now".