Avoid showing newly loaded dataset at wrong time
i.e. that the dataset is visible although another time is active on the
timeline.

Technically: Remove legacy code path after data loading. There was a
relic in Main._bgnd_open_paths_finish() which updated the legacy
DataLayerCollection and LayerStackTreeViewModel for the new dataset and
triggered animation updates based on the legacy Document interface.

Fixes #135