Ensure auto-update timer restart with granule merging
When a granule is merged into the data array of a previously
instantiated dataset no didAddDataset signal is sent. Currently this
signal is required though to trigger AutoUpdateManager.on_loading_done()
where the timer for the next update cycle is restarted. As a consequence
the auto-update halts when a granule is merged for the first time.

Fix this by replacing this signal route via a new signal
'didFinishLoading' that is unconditionally (both in case of success and
failure) emitted when loading is finished in the background thread and
which is connected to AutoUpdateManager.on_loading_done().