Ensure the correct datasets are removed properly
in Auto-Update mode the datasets from the previously loaded time(s) have
to be removed. The simple approach of just removing all datasets and
then load the current ones works without the granule merging; granules
removed by that process just would be loaded again.

With granule merging on (which is unconditionally on for Auto-Update
mode at the moment) this approach doesn't work, because the dataset
which the new granule(s) have been merged into must be kept and all
outdated ones be removed.

Thus: pass the uuid(s) of the currently loaded or updated dataset(s) to
the method on_loading_finished() via the signal didFinishLoading, filter
them out of the uuids of all datasets currently in memory and remove the
rest.

Furthermore with the new layer manager design it is not sufficient to
remove datasets via Document only, removing them must involve the
LayerModel. Implement this including the removal and purge of datasets
from Document and Workspace.

Implementation details:
- To implement the approach depicted above the LayerModel must know the
  Document, thus: pass it via the constructor, take Workspace from the
  passed Document
- Implement methods to find and remove/delete a dataset given by its
  uuid in/from the LayerModel no matter which LayerItem it belongs to.