Ensure display update after merging new granule(s)
The canvas needs to redraw after a granule has been injected into the
data array of a image dataset. It is unclear why this must be triggered
explicitly by calling canvas.update() and why calling set_data() of the
visual node in SceneGraphManager.update_basic_dataset() is not
sufficient to mark the node as modified which at other cases seems to
trigger redraw (by not understood VisPy magic).

The problem of this commit only becomes apparent (again) with the
previous commit

ad0383e6 ("Ensure display update after merging new granule(s)", 2022-04-07)

in place because the code path it removed triggered the redraw in a
different way.

Fixes #117