Fix hidden RGB layer with tiled geolocated image rendering
With 'display.use_tiled_geolocated_images: True' for RGB Composites the
according 'RGBCompositeLayer' visual node is set up with an additonal
z-transform of -50, a relic of how layers were sorted to different depths
with the legacy layer system. With the new layer system setup of the
scene-graph the visual nodes themselves are not moved but only the layer
group nodes they are childs of in the scene graph.

The wrong -50 shift in z direction just puts the 'RGBCompositeLayer'
right behind all other layers which explains, that it only can become
visible if all other layers are switched invisible.

Fix this by simply removing the code line that caused it.

Fixes #112