Revert "Show timeline even in auto update mode (for devel only)"
Showing the timeline also in Auto-Update mode was intentionally ony done
for development to see, which timesteps are currently in the LayerModel.

At at least as it was done it is not suitable for end users and it is
not an agreed feature yet.

TODO: Consider showing the timeline in Auto-Update mode as well, see #137.

This reverts commit 7096a89b4fecea12fc5b5d0cdb1aceba818c9b67.