Fix manual update of point probe display in graphs
such that clicking the button "Update" in the "Area Probe Graphs" pane
when "Auto Update" is off also updates/displays the red bar for the
probe value if one is active.

Implementation detail:
the call to update_point_probe_graph() is wrapped by a method
_update_point_probe_graph(), which queries the state and xy-position of
the active probe (currently only the default one), which are not
available in the signal emitted via the "Update" button click.

Fixes #141