Remove the visible/invisible separation of datasets for retiling
At the moment it seems that the implemented separation of datasets into
visible and invisible datasets has no effect on the retiling process.

This should prioritise the visible datasets for retiling, but the update
of the Scene graph is only triggered once.  Therefore, this
prioritisation is removed for now.

It may need to be reversed or re-implemented later to improve the
retiling times (see also #82).