Replace the saved value of Document._info_by_uuid with a dict
In Document._info_by_uuid `DocBasicDataset`s are stored as a value
and the corresponding UUID as a key. With the overloaded
layer management system and the code clean-up performed so far,
it is sufficient to store dicts as values in Document._info_by_uuid.
Thus, the DocBasicDataset is now replaced in Document._info_by_uuid
and the associated code is adapted accordingly.

Type annotations are adapted and unused imports are removed accordingly.