Refactor out a method to update the current timebase
The according functionality was buried in method to change the timebase,
which always triggers further actions in the time transformer policy
that have to be performed when the timebase changes.

To enable updating the current timebase without triggering the change
timebase action cascade, refactor part of
TimeTransformer.change_timebase() into own method.