Explicitly state the data 'kind' in some function calls
where the default kind IMAGE is wrong for retrieving content or product
data for multi-channel images from the workspace. Multi-channel images
have a slightly different Content class and without explicitly stating,
that a multichannel image is involved, the workspace would return data
as if it was from a single-channel image and this would cause issues.

It would be better if the methods
  BaseWorkspace._overview_content_for_uuid(),
  BaseWorkspace.get_content and
  BaseWorkspace._product_overview_content()
did not have a default value for the kind.
However, this will have to be adjusted with another refactoring
in the future. This is because the BaseWorkspace.get_content() method
is used in various places where no information about the type
of ProductDataset etc. is available.

Concerns #91