Adjust chunk merging to work with multichannel composites
Some complex composites offered by Satpy do not work with merging. This
cannot be solved unless Satpy provides an API to to get the complete
dependency tree/list of required files, when a composite is created from
other composites.

For failing composites, log a message that this is because
"data_reading.merge_with_existing" is configured to "True".

The problem is that we have to filter the files required to create the
needed composite dataset chunks, but:

For some composites the prerequisites don't tell exactly which channel
is needed to create a complex RGB, for example.  In other cases, the
prerequisites only tell a float value (for the wavelength), but this is
not enough information to infer the needed (this issue only occurs when
different channels are stored in spearate files as e.g. with SEVIRI
HRIT).