Update Conda complete packages list to Python 3.10
And:
Add some Satpy reader dependencies to Conda package list to reduce
number of exceptions thrown when opening the OpenFileWizard:
- 'bottleneck'
- 'h5netcdf'
- 'hdf5plugin'
- 'pyhdf'
- 'python-eccodes'
- 'rioxarray'

Since it is from a different channel, the package 'public_decomp_wt'
is commented out.

NOTE:
This is reach out for 'reproducible builds' but more work is needed to
get this done and keep it maintainable: it is not yet clear how to keep
both maintainability (allow for updates) and reproducibility.
The underlying problem is quite well depicted here:
https://pythonspeed.com/articles/conda-dependency-management/

TODO:
- Follow-up: Do we need different pkgs lists for Python 3.8/3.9./3.10?
  Also different lists for different platforms, or is it sufficient to
  remove the package version hashes in the list?
- There may be still a few Satpy or EUMETSAT's local_readers readers
  which throw execptions, could be because of missing/removed API in
  Satpy itself.
  The broken import tries to import 'satpy.readers.slstr_l2'.
- 'pygac' is not available on the channels 'conda-forge' or 'eumetsat'
  but 'bucricket', but it's another issue to decide about and add that
  channel if so.