Fix missing timeline manager ui on Windows
On Windows it may happen, that the timeline ui does not display
correctly. It can be observed that timeline.qml cannot be loaded and the
error
"c:/..../timeline.qml: Network error" is reported.

Fix the problem by explicitly telling the QML loader to interpret the
given file path as a local file path and not - the heuristic that
recognizes 'c:' as a protocol specifier is wrong here - as a network
URL.