Make zipped PyInstaller package with hidden internals
PyInstaller produces a directory with around 400 entries with the
mtgsift executable somewhere within.

Since this is not user friendly move all the pyinstaller output in a sub
directory 'mtgsift/lib' and prepare a 'bin' directory to add (later) a
start script for the user.

Finally create a zip archive from the result.