Fix undefined length of empty time transformer timeline
The timeline of a time transformer policy can be or become None (which could be
considered as equivalent to an empty List), while an animation is
running. In that case an exception

  TypeError: object of type 'NoneType' has no len()

could be thrown.

Avoid this by testing and returning 0 as length of timeline in case it
is None.

Concerns #47