Hide _GLGradientLineVisual._prepare_draw() from flake8 complexity check
... since refactoring this code would make future maintenance a
nightmare (even more): The class _GLGradientLineVisual is a modification
of a copy of the class 'GLLineVisual' from VisPy in
vispy/visuals/line/line.py around version v0.6.3. To be able to update
this implementation in the future to follow the development in VisPy
deviating from that code base should be avoided for now.

One thing which slightly simplifies the code and does updates it
minimally towards newer VisPy is to remove support for older VisPy
around the line

  self._program.vert["to_vec4"] = self._ensure_vec4_func(pos.shape[-1])

that was introduced with commit

2b3e39ac ("Fix _GLGradientLineVisual compatibility with newer versions of VisPy", 2022-04-12)