Handle algebraic layer colour limits (clims) correctly
Previously, a hard-coded value for the clims was used for
the algebraic layer.

With this commit, the invalid clims value is now set initially.

When the algebraic layer receives its first own records, this invalid
value is replaced with the correct clims value.  However, this only
happens once, when the algebraic layer has this invalid value for clims.

If the algebraic layer has no input layers, the clims value
becomes invalid again.