Revise chunk merging wrt Satpy RGB composites
i.e., fine tune the idea of commit
038af811 ("Adjust chunk merging to work with multichannel composites", 2022-09-20)

Since there should be only one place in the code where a default
value for a config key (here: "data_reading.merge_with_existing") is
determined, don't duplicate what is done already in Main.open_wizard().
Furthermore it is not that important, whether "merge_with_existing" was
*configured* but whether it actually could happen in the current import
pass, which is not the case when resampling is done.

Thus: Make determination of "merge_with_existing"-status consistent.