Make aImporter.from_product() code more comprehensible
... by renaming some methods and variables as follows:
- Methods:
  - _get_requires() -> _get_types_of_required_aux_files()
  - _required_files_set() -> _get_paths_of_required_aux_files()
  - _wanted_paths() -> _get_paths_in_scene_contributing_to_ds()
- Variables/parameters:
  - fn -> path (it's the loop variable of a loop over 'paths')
  - name -> ds_name ('ds' is a common abbreviation for dataset in the code)
  - new_filenames -> new_paths
  - required_files -> required_aux_files_paths
  - requires -> required_aux_file_types
  - wanted_paths -> contributing_paths

Amongst others this should lead to fewer ambiguities such as:
Why is the list of new 'paths' called 'new_*filenames*' and not
'new_paths'? It is better to stick to one way of naming things
and not mix synonyms.

This commit does not change the semantics,
It merely introduces renaming.