Simplify aImporter.from_product() for use with Satpy only
Since commit
482fa4ad ("Delete the no longer used GeoTiffImporter and GoesRPUGImporter classes...", 2022-08-31)

the only Product import code path is via Satpy readers / SatpyImporter,
thus we can safely assume (and use assertions for now to make sure we
assume correctly) that certain data structures are available and filled.
For example, there must be a (Satpy) Scene, which contains the same list
of file paths as referenced by the resources of the Product in process.

This allows us to wave some runtime checks and remove the according 'if'
statements and the associated nesting.