Make complex composites work a little better
... when 'data_readers.merge_with_existing: True' is configured.

The following case now works: A complex RGB composite - i.e., one, which
has prerequisites which are itself RGB composites - should be loaded and
there is no existing dataset for its sched_time yet - in this case, even
if 'data_readers.merge_with_existing' is True, no merging happens, so
the dataset can be loaded.

Only if merging is actually tried for such an RGB composite it cannot
succeed (as before): for the composite dependencies, the list of files
that must be present in the Scene to perform the loading of the data
chunks is not yet available in the Scene, because these are only
resolved during Scene.load().