Fix problems with loading categorical datasets: Provide min/max
... (aka "actual range") for categorical datasets, i.e. those with
"flag_values"/"flag_meaning" or "flag_masks": a pseudo range is
calculated.

Without such a fallback, there would be no way to automatically provide
color mapping, and because of the way the loading and setup process is
structured, loading of categorical data for which no default color
mapping is defined would be completely broken.

This improves commit
a3f56ac5 ("Enable to get the actual range values for a dataset", 2022-11-24)

Concerns #48