Set default colormap to 'viridis' instead of 'gray'
The default color map is used for 'unknown' datasets, i.e. those for
which no color map is configured or not implemented in the 'Guidebook'.

The 'viridis' colormap is a gradient from dark purple to blue to green
and finally yellow. It seems to be more suitable for recognizing
features at first glance than the gray colormap.

Concerns #48