Revise handling of missing 'valid_range' globally
Commit
bb5fa55c ("Overwork the calculation of the valid range...", 2022-11-24)
addressed the issue that some product types do not specify a 'valid
range' in their metadata. This was necessary because without knowing the
range in which the data lies, there is no way to automatically adjust a
color mapping. In said commit, the approach that existed in sift before
- not clearly distinguishing between valid range and color boundaries -
was continued and where the valid range was not available, the so-called
'actual range' (the pair of values given by the minimum and maximum
value in the data as computed by the statistics module) was used to
define a valid range.

A disadvantage of that approach is, that after setting the valid range
from the actual range there is no way any more to know, where the valid
range came from.

This is considered problematic in the long term, it may cause trouble in
the future and increase maintenance costs.

Thus with this commit partially revert commit
bb5fa55c ("Overwork the calculation of the valid range...", 2022-11-24)
and take another approach:

For all data it can be assumed that an 'actual_range' is available,
because it can be computed from the data (even for the special case of
categorial data providing flag_values/flag_meanings or flag_masks, but
this is not right now adressed by this but by a future commit).

All code using the valid_range should be enabled to deal with the fact
that it may be missing and fall back to the actual_range or, if this
also fails for whatever reason, use a predefined FALLBACK_RANGE.

Most functionality should work for data like "radiance" and "counts",
which don't come with a valid range, and as well for algebraics, but
further adjustments are provided by the next commits.

Concerns #48, #132

Fixes #154, #155