Let OpenFileWizard restart at first page in previous state
i.e., without reinitializing page 1: It is only initialized once when
the wizard is created and opened the first time.

As a result, the user sees the first page in the same state as when she
left it. Only the second page of the wizard is reinitialised when it is
accessed, which in principle cannot be completely avoided. The state can
only be partially maintained. This will be dealt with in the next
commit.