Let OpenFileWizard try to restore resampler
at the second wizard page. The resampler only changes, if the previously
chosen is known to not work for the data in the files selected at page
one: not all resamplers work for data with a SwathDefinition.

TODO: Not all non-functioning combinations are recognised, in fact they
are not even all known.