Make raised dock widgets appear first in their tabs group
It is more intuitive that the dock widgets that are visible at startup
have the first tab position within their tab group.

Implement the tab grouping and raising of the widgets to the top
accordingly.

Implementation Detail: Since Qt does not provide an API to control the
order of dock widgets within a docking area, the order in which the
widgets are created is important. To keep the Layer Manager widget at
the bottom, its XML element must be moved down in the .ui file.

By the way, also fix a problem with the implementation of commit
  533fe068 ("Make buttons 'Pan/Zoom', 'Point', 'Region' exclusive", 2022-04-04)
and
  6798bea2 ("Revert partial "Make buttons 'Pan/Zoom', 'Point', 'Region' exclusive", 2022-08-01),
which result in a corrupted .ui file uwsift/ui/pov_main.ui. As a result,
the QButtonGroup responsible for grouping said buttons into an exclusive
group is lost when editing the file with Qt Designer.

With this commit, the button group is created as it should be (with Qt
Designer), resulting in a consistent .ui file.