Fix display of data resolution for data with `SwathDefinition`
SwathDefinitions have no 'pixel_size_x' or 'pixel_size_y' attributes
that are displayed as resolution for AreaDefinition data. On the other
hand, for data with SwathDefinition there is a 'resolution' attribute in
the corresponding metadata (info).

Display the value of the latter as resolution, but not like e.g.
'1.0 km / 1.0 km' as for AreaDefinition data, but only one-dimensional
like e.g. '1.0 km': We only have one value, so we should not pretend to
have two.

Concerns #148

Fixes #156