Revise default colormaps configuration
Rename the settings
- colormap_name -> colormap
- climits -> range

The latter is changed to express, that the tuple used for setting the
range for the colour mapping currently actually defines the "valid range"
internally, when it is not provided by the data. In fact, the setting in
the configuration even overwrites the valid range of the data.

So the setting should be used with caution and should be used only for
data with the valid range not set, but if you really want, you have
the possibility to overwrite the valid range with the configuration.

TODO: Revise the configuraton and enable to set the valid range for a
product in the readers configuration. Then the according setting in the
colormap configuration may be removed or renamed to 'climits' again.

Convert configured value for range to a tuple

If the value for range is set like a tuple, it is interpreted
as a string.
Therefore, the range must be defined as a list in the configuration.
In the application, however, a tuple is expected later.

Concerns #48