Avoid flood of signals at OpenFileWizard page initialization
While the 2nd wizard page "Product Selection" is initialized and the table
of available products is filled it doesn't make sense to check, whether
the page already is complete (in the QWizardPage sense). With the current
signal slot setup, a needless check is performed for each table cell
(QTableWidgetItem) added.

To avoid this, temporarily disable the signal slot connection that triggers
such a check during the initialization of the wizard page.

Note: The behavior that this commit fixes is not normally noticeable to
the user, but was extremely annoying during a debugging session that
dealt with the affected code.