Let file dialog open in previous state
When the file dialog for selecting a data directory is opened from
within the OpenFileWizard, it starts with the previously selected
folder, which, however, only insufficiently corresponds to the state in
which the dialog was left the last time.

Especially in combination with the directory tree navigation introduced
with commit
0e35301d ("Enable tree navigation in "Select Data Directory" dialog", 2022-12-07)
this feels unnatural.

Improve the situation by reusing the file dialog instance. This way, it
will automatically reopen in the very same state in which it was left.