Make sure no color editing is offered for empty layers
The color editing part of the Layer Details pane is not cleared if a
IMAGE layer becomes empty (again).

Currently (as long as there is no way to remove single datasets) this
can only happen for Algebraic layers, when its inputs are unset (either
explicitly by the user or implicitly because an input layer is removed).

For such a layer there is no sensible way to determine a valid_range or
an actual_range which is required to offer color range editing.
Therefore the corresponding user interface should be hidden from the
user in this situation.

Fix this by restructuring the update process.