Make sure current projection is set for resampling
in the "Select Products" page of the OpenFileWizard.

Before commit
6b1c3ed1 ("Reuse the OpenFileWizard instance", 2021-02-22)
a new OpenFileWizard instance was created each time, so setting the
current projection in its constructor was sufficient. It has been
overseen, that an explicit update has to be performed when the
OpenFileWizard instance is reused.

Implementation detail: the OpenFileWizard.ui.projectionComboBox now
shares the model of Main.ui.projectionComboBox to make sure, the index
used to select the projection, Document.current_projection_index(),
refers to the same projection. Otherwise there is no guarantee by
design, that the two ComboBoxes are consistent to each other.

Fixes #157