Fix internal RGBs failing due to new clims treatment
In the commits revising how the layers details panel retrieves settings
from color limits (climits) stored in a layers Presentation

85b26f8c ("Revise color range handling in Layer Details pane (2)", 2022-12-09)
1e770741 ("Revise handling of missing 'valid_range' globally", 2022-12-01)

it was overseen, that for the internal recipe RGB composites climits is a
list of pairs (i.e. [(r_min, r_max), (g_min, g_max), (b_min, b_max)])
from which the sequence extraction of two values fails.

Deal with this by ignoring the climits in that case, because there won't
be color range sliders in the Layers Details pane anyway.