Revise BaseWorkspace.get_min_max_value_for_dataset_by_uuid()
Change log message from level ERROR to DEBUG because the notified case
happens regularly for Satpy RGB Composites: Statistics are not
calculated for them on purpose, see commit
b4ecc963 ("Exclude Satpy RGB Composites from statistics calculation", 2022-12-01)

When inspecting the code for this change it was noticed, that the
implementation has an issue: The method doesn't return anything when the
input parameter 'uuid' is None, but doesn't fail, the caller then has to
deal with that it might get a None result. It is considered better
practice to fail early in the method with an exception and return at
least (None, None) when executing normally.

Change the implementatin accordingly.

Note: The assertions in the code do not change the behaviour of the
method, they are only there to explitly state under which conditions the
method will raise exceptions.